data/2025/stego_flag/requirements.txt
pillow
torch
numpy